minten

+++ contracts/interfaces/IFarmtroller.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.0;
 
 interface IFarmtroller {
 
-    function burningToken(address _address) external returns(uint);
+    function burningToken(address _address) external returns(uint256);
 }
